Adds implementation and support for Attachment option type in interactions.
